requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.49
python-dotenv==1.2.2
Werkzeug==3.0.0
Jinja2==3.1.2
MarkupSafe==2.1.1
click==8.1.7
itsdangerous==2.1.2
